requirements.txt

Pillow==10.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0  # для локальной разработки